Make generated visual artifacts portable and complete
